+++ src/plonemeeting/restapi/serializer/dxfields.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from imio.helpers.content import get_schema_fields
 from plone.dexterity.interfaces import IDexterityContent
 from plone.restapi.interfaces import IFieldSerializer
 from plone.restapi.serializer.converters import json_compatible
@@ -13,6 +14,7 @@ from zope.schema.interfaces import IChoice
 from zope.schema.interfaces import ICollection
 from zope.schema.interfaces import IField
 from zope.schema.interfaces import IVocabularyTokenized
+from zope.schema.vocabulary import SimpleTerm
 
 
 @adapter(IChoice, IDexterityContent, IPMRestapiLayer)
@@ -30,11 +32,22 @@ class PMChoiceFieldSerializer(ChoiceFieldSerializer):
         if value is not None and IVocabularyTokenized.providedBy(vocab):
             try:
                 term = self.field.vocabulary.getTerm(value)
-                value = {"token": term.token, "title": term.title}
-            # Some fields (e.g. language) have a default value that is not in
-            # vocabulary
+            # If not found, get it from the z3c.form that maybe uses a MissingTerms adapter...
             except LookupError:
-                pass
+                try:
+                    view = self.context.restrictedTraverse("@@view")
+                    view.update()
+                    # widget_name is the field name prefixed with behavior if any
+                    widget_name = [field for field in get_schema_fields(self.context, prefix=True)
+                                   if field[1].__name__ == self.field.__name__][0][0]
+                    widget = view.widgets[widget_name]
+                    term = widget.terms.getTerm(value)
+                except Exception:
+                    # at worse use value as title
+                    term = SimpleTerm(value, title=value)
+            finally:
+                value = {"token": term.token, "title": term.title}
+
         return json_compatible(value)